// src/payload/payload-types.ts
export type CartItems =
  | {
      product?: (number | null) | Product;
      quantity?: number | null;
      id?: string | null;
    }[]
  | null;
// ...
export interface Order {
  id: number;
  orderedBy?: (number | null) | User;
  stripePaymentIntentID?: string | null;
  total: number;
  items?:
    | {
        product: number | Product;
        price?: number | null;
        quantity?: number | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
export interface User {
  id: number;
  name?: string | null;
  roles?: ('admin' | 'customer')[] | null;
  purchases?: (number | Product)[] | null;
  stripeCustomerID?: string | null;
  cart?: {
    items?: CartItems;
  };
  skipSync?: boolean | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password: string | null;
}